# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "思维模型平台 - 基于结构化思维模型和AI的复杂问题深度思考平台",
        "title": "Thinking Models Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@thinkingmodels.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:2500",
    "basePath": "/api/v1",
    "paths": {
...
        "/subject/topic": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的思考课题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课题管理"
                ],
                "summary": "创建课题",
                "parameters": [
                    {
                        "description": "创建课题请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/thinkingModels_domain_subject_topic.CreateTopic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_subject_topic.TopicInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/subject/topic/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页查询课题列表，支持多种筛选条件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课题管理"
                ],
                "summary": "查询课题列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标题模糊查询",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态：0-进行中，1-已完成，2-已归档",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "优先级：1-低，2-中，3-高",
                        "name": "priority",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_subject_topic.ListTopicResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/subject/topic/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据ID获取课题详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课题管理"
                ],
                "summary": "获取课题详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课题ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/thinkingModels_domain_subject_topic.TopicInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "课题不存在",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "trace_id": {}
            }
        },
        "thinkingModels_domain_subject_topic.CreateTopic": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "deadline": {
                    "description": "截止日期（ISO格式字符串）",
                    "type": "string"
                },
                "description": {
                    "description": "课题描述",
                    "type": "string",
                    "maxLength": 2000
                },
                "priority": {
                    "description": "优先级: 1=低, 2=中, 3=高",
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "tags": {
                    "description": "标签，逗号分隔",
                    "type": "string",
                    "maxLength": 500
                },
                "title": {
                    "description": "课题标题",
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
        "thinkingModels_domain_subject_topic.ListTopicResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/thinkingModels_domain_subject_topic.TopicInfo"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "thinkingModels_domain_subject_topic.TopicInfo": {
            "type": "object",
            "properties": {
                "completeTime": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modelId": {
                    "type": "integer"
                },
                "modelName": {
                    "description": "关联的模型名称",
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "priorityText": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "tagsList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token认证，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}